backend/requirements.txt
django==3.0.5
celery==4.4.6
psycopg2-binary==2.8.5
gunicorn==20.0.4
djangorestframework==3.11.0
coverage==5.2
requests==2.24.0